type change and first method

+++ include/abod.h
@@ -13,8 +13,8 @@ std::vector<listnd> randomProjection(std::vector<vecnd> dataset,
 		int t);
 
 vecnd firstMomentEstimator(std::vector<listnd> listset,
-		int t,long long n);
+		int t, size_t n);
 
 vecnd frobeniusNorm(std::vector<listnd> listset,
-		int t,long long n);
+		int t, size_t n);
 #endif
